CFI防护代码整改和重复日志整改

+++ frameworks/native/capi/screencapture/native_avscreen_capture.cpp
@@ -860,7 +860,7 @@ OH_AVSCREEN_CAPTURE_ErrCode OH_AVScreenCapture_ContentFilter_AddWindowContent(
                              AV_SCREEN_CAPTURE_ERR_INVALID_VAL, "input window invalid!");
     std::vector<uint64_t> vec;
     for (int32_t i = 0; i < windowCount; i++) {
-        if (static_cast<uint64_t>(*(windowIDs + i)) >= 0) {
+        if (static_cast<int32_t>(*(windowIDs + i)) >= 0) {
             vec.push_back(static_cast<uint64_t>(*(windowIDs + i)));
         }
     }
